Change the way of indexing the VMs
Since whenever we insert a charater that is not a number we got result 0
for cin >> (size_t)selected. Because of 0 was the index of the first VM
every character that is not a number could start the first VM.

To resolve this problem we changed the index-starting from 0 to 1.